Client secret now configurable #506
- client secret can now be set by
 "checkmarx.clientSecret" parameter. When not set
 default is used
- removed completele from checkmarx install setup
- created test
