+++ sechub-adapter-checkmarx/src/main/java/com/daimler/sechub/adapter/checkmarx/CheckmarxConfig.java
@@ -8,15 +8,7 @@ import com.daimler.sechub.adapter.AbstractCodeScanAdapterConfigBuilder;
 
 public class CheckmarxConfig extends AbstractCodeScanAdapterConfig implements CheckmarxAdapterConfig {
 
-    /**
-     * This is the "client secret" is listed at <a href=
-     * "https://checkmarx.atlassian.net/wiki/spaces/KC/pages/1187774721/Using+the+CxSAST+REST+API+v8.6.0+and+up"
-     * >public Checkmarx documentation</a>
-     *
-     * Being not really a secret but just a visible constant in public space it's
-     * okay to contain this inside code.
-     */
-    public static final String DEFAULT_CLIENT_SECRET = "014DF517-39D1-4453-B7B3-9930C563627C";
+    public static final String DEFAULT_CLIENT_SECRET = CheckmarxConstants.DEFAULT_CLIENT_SECRET;
 
     private String teamIdForNewProjects;
     private InputStream sourceCodeZipFileInputStream;
