+++ sechub-scan-product-checkmarx/src/main/java/com/daimler/sechub/domain/scan/product/checkmarx/CheckmarxInstallSetup.java
@@ -5,12 +5,5 @@ import com.daimler.sechub.domain.scan.InstallSetup;
 
 public interface CheckmarxInstallSetup extends InstallSetup{
 
-    /**
-     * Get 'client secret' value
-     * 
-     * @return
-     */
-    public String getClientSecret();
-
     public boolean isHavingUntrustedCertificate();
 }
